♻️ Move styles into styles files
